Separate ISigner and IVerifier interfaces.
