+++ bftengine/include/bftengine/IKeyExchanger.hpp
@@ -10,7 +10,7 @@
 // file.
 
 #pragma once
-#include "crypto_utils.hpp"
+#include "cryptopp_utils.hpp"
 #include <string>
 #include <cstdint>
