+++ bftengine/tests/clientsManager/ClientsManager_test.cpp
@@ -24,9 +24,9 @@ using bftEngine::impl::SigManager;
 using bftEngine::ReplicaConfig;
 using bftEngine::ReservedPagesClientBase;
 using bftEngine::test::ReservedPagesMock;
-using concord::util::crypto::Crypto;
+using concord::util::cryptopp_utils::Crypto;
 using concord::util::crypto::KeyFormat;
-using concord::util::crypto::RSASigner;
+using concord::util::cryptopp_utils::RSASigner;
 using concord::secretsmanager::ISecretsManagerImpl;
 using concordUtil::Timers;
 using std::chrono::milliseconds;
