+++ client/reconfiguration/src/default_handlers.cpp
@@ -64,7 +64,7 @@ void ClientTlsKeyExchangeHandler::exchangeTlsKeys(const std::string& pkey_path,
                                                   const std::string& cert_path,
                                                   const uint64_t blockid) {
   // Generate new key pair
-  auto new_cert_keys = concord::util::crypto::Crypto::instance().generateECDSAKeyPair(
+  auto new_cert_keys = concord::util::cryptopp_utils::Crypto::instance().generateECDSAKeyPair(
       concord::util::crypto::KeyFormat::PemFormat, concord::util::crypto::CurveType::secp384r1);
 
   std::string master_key = sm_->decryptFile(master_key_path_).value_or(std::string());
@@ -151,9 +151,9 @@ bool ClientMasterKeyExchangeHandler::validate(const State& state) const {
 bool ClientMasterKeyExchangeHandler::execute(const State& state, WriteState& out) {
   LOG_INFO(getLogger(), "execute transaction signing key exchange request");
   // Generate new key pair
-  auto hex_keys = concord::util::crypto::Crypto::instance().generateRsaKeyPair(
+  auto hex_keys = concord::util::cryptopp_utils::Crypto::instance().generateRsaKeyPair(
       2048, concord::util::crypto::KeyFormat::HexaDecimalStrippedFormat);
-  auto pem_keys = concord::util::crypto::Crypto::instance().RsaHexToPem(hex_keys);
+  auto pem_keys = concord::util::cryptopp_utils::Crypto::instance().RsaHexToPem(hex_keys);
 
   concord::messages::ReconfigurationRequest rreq;
   concord::messages::ClientExchangePublicKey creq;
