+++ kvbc/src/pruning_handler.cpp
@@ -29,14 +29,14 @@ void RSAPruningSigner::sign(concord::messages::LatestPrunableBlock& block) {
 }
 
 RSAPruningSigner::RSAPruningSigner(const std::string& key)
-    : signer_{std::make_unique<concord::util::crypto::RSASigner>(
+    : signer_{std::make_unique<concord::util::cryptopp_utils::RSASigner>(
           key, concord::util::crypto::KeyFormat::HexaDecimalStrippedFormat)} {}
 
 RSAPruningVerifier::RSAPruningVerifier(const std::set<std::pair<uint16_t, const std::string>>& replicasPublicKeys) {
   auto i = 0u;
   for (auto& [idx, pkey] : replicasPublicKeys) {
     replicas_.push_back(Replica{idx,
-                                std::make_unique<concord::util::crypto::RSAVerifier>(
+                                std::make_unique<concord::util::cryptopp_utils::RSAVerifier>(
                                     pkey, concord::util::crypto::KeyFormat::HexaDecimalStrippedFormat)});
     const auto ins_res = replica_ids_.insert(replicas_.back().principal_id);
     if (!ins_res.second) {
