+++ reconfiguration/include/reconfiguration/reconfiguration_handler.hpp
@@ -18,7 +18,7 @@
 #include "ireconfiguration.hpp"
 #include "OpenTracing.hpp"
 #include "SigManager.hpp"
-#include "crypto_utils.hpp"
+#include "cryptopp_utils.hpp"
 
 namespace concord::reconfiguration {
 class BftReconfigurationHandler : public IReconfigurationHandler {
@@ -26,7 +26,7 @@ class BftReconfigurationHandler : public IReconfigurationHandler {
   BftReconfigurationHandler();
   bool verifySignature(uint32_t sender_id, const std::string &data, const std::string &signature) const override;
 
-  std::unique_ptr<concord::util::crypto::IVerifier> verifier_;
+  std::unique_ptr<concord::util::cryptointerface::IVerifier> verifier_;
 };
 class ReconfigurationHandler : public BftReconfigurationHandler {
  public:
